requirements.txt
tensorflow
opencv-python
gdown
numpy
Pillow